requirements.txt
streamlit
langchain-openai
langchain-core
httpx
pandas
numpy
pillow
python-dotenv
sqlalchemy